pyproject.toml
[tool.poetry]
name = "onecontainer-cloud-tool"
version = "0.2.0"
description = "The oneContainer-cloud-tool utility helps to deploy containers to public cloud."
authors = ["Duran Padilla, Mauricio <mauricio.duran.padilla@intel.com>", "Rahul Unnikrishnan Nair <rahul.unnikrishnan.nair@intel.com>"]

[tool.poetry.dependencies]
python = "3.7"
cryptography= "3.3.2"
boto3 = "1.17.20"
click = "^7.1.2"
loguru = "^0.5.3"
rich = "^9.6.2"
DateTime = "^4.3"
requests = "^2.25.1"
azure-identity = "1.5.0"
azure-mgmt-resource = "^15.0.0"
azure-mgmt-compute = "^18.1.0"
azure-mgmt-network = "^17.1.0"
pycryptodomex = "^3.9.9"
pylint = "^2.7.2"
pywin32 = {version = "^300", platform = "windows"}
google-api-python-client = "*"
google-auth = "*"
progress = "^1.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
icecream = {version = "^2.1.0", allow-prereleases = true}
black = "^20.8b1"
pytest-cov = "^2.11.1"
ipdb = "^0.13.9"

[tool.poetry.scripts]
onecontainer-cloud-tool = "onecontainer_cloud_tool.cli:cli"
oneContainer-Cloud-Tool = "onecontainer_cloud_tool.cli:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

